Restructure and extract constants
